Implement usefull new basic tree_cos LCS struct
- Implement usefull new basic
 tree_closest_offset_sum LCS struct.

[Issue: #X]

+++ src/lib.rs
@@ -3,7 +3,7 @@ pub mod new_algos;
 pub mod slow_lcs;
 pub mod lcs_trait;
 
-pub type MaxInputLengthType = u8;
+pub type MaxInputLengthType = usize;
 pub const UNIQUE_VALUES: usize = 256;
 
 #[cfg(test)]
